Improve CI: add Miri tests as postsubmit
* Comment coverage report on PR and push.
* Fix Miri test failures.
  * Manually generate VkLayerInstanceCreateInfo to fix a Miri failure with stack borrow.
  * Handle null `pp_enabled_extension_names`.
  * Remove 3 test failures for Miri to check later.
* Add Miri tests as postsubmit.
* Use nextest config file to manage CI config.
