+++ .config/nextest.toml
@@ -1,0 +1,7 @@
+[profile.default]
+retries = 0
+test-threads = "num-cpus"
+
+[profile.ci]
+failure-output = "immediate-final"
+fail-fast = false
