+++ vulkan-layer/src/bindings/vk_layer.rs
@@ -17,6 +17,8 @@
 #![allow(non_upper_case_globals)]
 #![allow(dead_code)]
 
+use std::mem::MaybeUninit;
+
 use ash::vk;
 
 pub mod generated;
@@ -36,6 +38,40 @@ pub struct VkLoaderFeatureFlags(vk::Flags);
 ash::vk_bitflags_wrapped!(VkLoaderFeatureFlags, vk::Flags);
 impl VkLoaderFeatureFlags {
     pub const _PHYSICAL_DEVICE_SORTING: Self = Self(0x00000001);
+}
+
+#[repr(C)]
+#[derive(Debug, Default, Copy, Clone)]
+pub struct VkLayerInstanceCreateInfoUFieldLayerDeviceField {
+    pub pfnLayerCreateDevice: PFN_vkLayerCreateDevice,
+    pub pfnLayerDestroyDevice: PFN_vkLayerDestroyDevice,
+}
+
+#[repr(C)]
+#[derive(Copy, Clone)]
+pub union VkLayerInstanceCreateInfoUField {
+    pub pLayerInfo: *mut VkLayerInstanceLink,
+    pub pfnSetInstanceLoaderData: PFN_vkSetInstanceLoaderData,
+    pub layerDevice: VkLayerInstanceCreateInfoUFieldLayerDeviceField,
+    pub loaderFeatures: VkLoaderFeatureFlags,
+}
+
+impl Default for VkLayerInstanceCreateInfoUField {
+    fn default() -> Self {
+        let mut s = MaybeUninit::<Self>::uninit();
+        unsafe {
+            std::ptr::write_bytes(s.as_mut_ptr(), 0, 1);
+            s.assume_init()
+        }
+    }
+}
+
+#[repr(C)]
+pub struct VkLayerInstanceCreateInfo {
+    pub sType: VkStructureType,
+    pub pNext: *const ::std::os::raw::c_void,
+    pub function: VkLayerFunction,
+    pub u: VkLayerInstanceCreateInfoUField,
 }
 
 unsafe impl vk::TaggedStructure for VkLayerInstanceCreateInfo {
