+++ vulkan-layer/src/bindings/vk_layer/generated.rs
@@ -25,49 +25,6 @@ pub use windows::*;
 
 /* automatically generated by rust-bindgen 0.65.1 */
 
-#[repr(C)]
-pub struct __BindgenUnionField<T>(::std::marker::PhantomData<T>);
-impl<T> __BindgenUnionField<T> {
-    #[inline]
-    pub const fn new() -> Self {
-        __BindgenUnionField(::std::marker::PhantomData)
-    }
-    #[inline]
-    pub unsafe fn as_ref(&self) -> &T {
-        ::std::mem::transmute(self)
-    }
-    #[inline]
-    pub unsafe fn as_mut(&mut self) -> &mut T {
-        ::std::mem::transmute(self)
-    }
-}
-impl<T> ::std::default::Default for __BindgenUnionField<T> {
-    #[inline]
-    fn default() -> Self {
-        Self::new()
-    }
-}
-impl<T> ::std::clone::Clone for __BindgenUnionField<T> {
-    #[inline]
-    fn clone(&self) -> Self {
-        Self::new()
-    }
-}
-impl<T> ::std::marker::Copy for __BindgenUnionField<T> {}
-impl<T> ::std::fmt::Debug for __BindgenUnionField<T> {
-    fn fmt(&self, fmt: &mut ::std::fmt::Formatter<'_>) -> ::std::fmt::Result {
-        fmt.write_str("__BindgenUnionField")
-    }
-}
-impl<T> ::std::hash::Hash for __BindgenUnionField<T> {
-    fn hash<H: ::std::hash::Hasher>(&self, _state: &mut H) {}
-}
-impl<T> ::std::cmp::PartialEq for __BindgenUnionField<T> {
-    fn eq(&self, _other: &__BindgenUnionField<T>) -> bool {
-        true
-    }
-}
-impl<T> ::std::cmp::Eq for __BindgenUnionField<T> {}
 pub type PFN_GetPhysicalDeviceProcAddr = ::std::option::Option<
     unsafe extern "C" fn(
         instance: VkInstance,
@@ -257,205 +214,6 @@ pub type PFN_vkLayerDestroyDevice = ::std::option::Option<
         destroyFunction: PFN_vkDestroyDevice,
     ),
 >;
-#[repr(C)]
-pub struct VkLayerInstanceCreateInfo {
-    pub sType: VkStructureType,
-    pub pNext: *const ::std::os::raw::c_void,
-    pub function: VkLayerFunction,
-    pub u: VkLayerInstanceCreateInfo__bindgen_ty_1,
-}
-#[repr(C)]
-pub struct VkLayerInstanceCreateInfo__bindgen_ty_1 {
-    pub pLayerInfo: __BindgenUnionField<*mut VkLayerInstanceLink>,
-    pub pfnSetInstanceLoaderData: __BindgenUnionField<PFN_vkSetInstanceLoaderData>,
-    pub layerDevice: __BindgenUnionField<VkLayerInstanceCreateInfo__bindgen_ty_1__bindgen_ty_1>,
-    pub loaderFeatures: __BindgenUnionField<VkLoaderFeatureFlags>,
-    pub bindgen_union_field: [u64; 2usize],
-}
-#[repr(C)]
-#[derive(Debug, Default, Copy, Clone)]
-pub struct VkLayerInstanceCreateInfo__bindgen_ty_1__bindgen_ty_1 {
-    pub pfnLayerCreateDevice: PFN_vkLayerCreateDevice,
-    pub pfnLayerDestroyDevice: PFN_vkLayerDestroyDevice,
-}
-#[test]
-fn bindgen_test_layout_VkLayerInstanceCreateInfo__bindgen_ty_1__bindgen_ty_1() {
-    const UNINIT: ::std::mem::MaybeUninit<VkLayerInstanceCreateInfo__bindgen_ty_1__bindgen_ty_1> =
-        ::std::mem::MaybeUninit::uninit();
-    let ptr = UNINIT.as_ptr();
-    assert_eq!(
-        ::std::mem::size_of::<VkLayerInstanceCreateInfo__bindgen_ty_1__bindgen_ty_1>(),
-        16usize,
-        concat!(
-            "Size of: ",
-            stringify!(VkLayerInstanceCreateInfo__bindgen_ty_1__bindgen_ty_1)
-        )
-    );
-    assert_eq!(
-        ::std::mem::align_of::<VkLayerInstanceCreateInfo__bindgen_ty_1__bindgen_ty_1>(),
-        8usize,
-        concat!(
-            "Alignment of ",
-            stringify!(VkLayerInstanceCreateInfo__bindgen_ty_1__bindgen_ty_1)
-        )
-    );
-    assert_eq!(
-        unsafe { ::std::ptr::addr_of!((*ptr).pfnLayerCreateDevice) as usize - ptr as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(VkLayerInstanceCreateInfo__bindgen_ty_1__bindgen_ty_1),
-            "::",
-            stringify!(pfnLayerCreateDevice)
-        )
-    );
-    assert_eq!(
-        unsafe { ::std::ptr::addr_of!((*ptr).pfnLayerDestroyDevice) as usize - ptr as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(VkLayerInstanceCreateInfo__bindgen_ty_1__bindgen_ty_1),
-            "::",
-            stringify!(pfnLayerDestroyDevice)
-        )
-    );
-}
-#[test]
-fn bindgen_test_layout_VkLayerInstanceCreateInfo__bindgen_ty_1() {
-    const UNINIT: ::std::mem::MaybeUninit<VkLayerInstanceCreateInfo__bindgen_ty_1> =
-        ::std::mem::MaybeUninit::uninit();
-    let ptr = UNINIT.as_ptr();
-    assert_eq!(
-        ::std::mem::size_of::<VkLayerInstanceCreateInfo__bindgen_ty_1>(),
-        16usize,
-        concat!(
-            "Size of: ",
-            stringify!(VkLayerInstanceCreateInfo__bindgen_ty_1)
-        )
-    );
-    assert_eq!(
-        ::std::mem::align_of::<VkLayerInstanceCreateInfo__bindgen_ty_1>(),
-        8usize,
-        concat!(
-            "Alignment of ",
-            stringify!(VkLayerInstanceCreateInfo__bindgen_ty_1)
-        )
-    );
-    assert_eq!(
-        unsafe { ::std::ptr::addr_of!((*ptr).pLayerInfo) as usize - ptr as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(VkLayerInstanceCreateInfo__bindgen_ty_1),
-            "::",
-            stringify!(pLayerInfo)
-        )
-    );
-    assert_eq!(
-        unsafe { ::std::ptr::addr_of!((*ptr).pfnSetInstanceLoaderData) as usize - ptr as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(VkLayerInstanceCreateInfo__bindgen_ty_1),
-            "::",
-            stringify!(pfnSetInstanceLoaderData)
-        )
-    );
-    assert_eq!(
-        unsafe { ::std::ptr::addr_of!((*ptr).layerDevice) as usize - ptr as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(VkLayerInstanceCreateInfo__bindgen_ty_1),
-            "::",
-            stringify!(layerDevice)
-        )
-    );
-    assert_eq!(
-        unsafe { ::std::ptr::addr_of!((*ptr).loaderFeatures) as usize - ptr as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(VkLayerInstanceCreateInfo__bindgen_ty_1),
-            "::",
-            stringify!(loaderFeatures)
-        )
-    );
-}
-impl Default for VkLayerInstanceCreateInfo__bindgen_ty_1 {
-    fn default() -> Self {
-        let mut s = ::std::mem::MaybeUninit::<Self>::uninit();
-        unsafe {
-            ::std::ptr::write_bytes(s.as_mut_ptr(), 0, 1);
-            s.assume_init()
-        }
-    }
-}
-#[test]
-fn bindgen_test_layout_VkLayerInstanceCreateInfo() {
-    const UNINIT: ::std::mem::MaybeUninit<VkLayerInstanceCreateInfo> =
-        ::std::mem::MaybeUninit::uninit();
-    let ptr = UNINIT.as_ptr();
-    assert_eq!(
-        ::std::mem::size_of::<VkLayerInstanceCreateInfo>(),
-        40usize,
-        concat!("Size of: ", stringify!(VkLayerInstanceCreateInfo))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<VkLayerInstanceCreateInfo>(),
-        8usize,
-        concat!("Alignment of ", stringify!(VkLayerInstanceCreateInfo))
-    );
-    assert_eq!(
-        unsafe { ::std::ptr::addr_of!((*ptr).sType) as usize - ptr as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(VkLayerInstanceCreateInfo),
-            "::",
-            stringify!(sType)
-        )
-    );
-    assert_eq!(
-        unsafe { ::std::ptr::addr_of!((*ptr).pNext) as usize - ptr as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(VkLayerInstanceCreateInfo),
-            "::",
-            stringify!(pNext)
-        )
-    );
-    assert_eq!(
-        unsafe { ::std::ptr::addr_of!((*ptr).function) as usize - ptr as usize },
-        16usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(VkLayerInstanceCreateInfo),
-            "::",
-            stringify!(function)
-        )
-    );
-    assert_eq!(
-        unsafe { ::std::ptr::addr_of!((*ptr).u) as usize - ptr as usize },
-        24usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(VkLayerInstanceCreateInfo),
-            "::",
-            stringify!(u)
-        )
-    );
-}
-impl Default for VkLayerInstanceCreateInfo {
-    fn default() -> Self {
-        let mut s = ::std::mem::MaybeUninit::<Self>::uninit();
-        unsafe {
-            ::std::ptr::write_bytes(s.as_mut_ptr(), 0, 1);
-            s.assume_init()
-        }
-    }
-}
 #[repr(C)]
 pub struct VkLayerDeviceLink_ {
     pub pNext: *mut VkLayerDeviceLink_,
